Add GameLive, GameView and a .leex template
We'll start out with rendering the board. First, we add a live view to
our application to handle rendering and updating the board. It's called
`GameLive` and it has a `render/1` and a `mount/2` callback function.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      use Phoenix.LiveView

      def render(assigns) do
        HayagoWeb.GameView.render("index.html", assigns)
      end

      def mount(_session, socket) do
        {:ok, assign(socket, state: %Hayago.State{})}
      end
    end

The `mount/2` callback is tasked with setting up the assings in the
`socket` to set the initial state for the view. We use it to create a
new state, and we add it to the socket assigns to make it available in
the template.

Next, we'll add a template to render the board. We'll name it
`index.html.leex`,to make it *Live EEx* template. Although similar to
regular EEx templates, the live EEx can track changes to send a minimal
amount of data over the wire whenever the view updates.

    # lib/hayago_web/templates/game/index.html.leex
    <div class="board <%= @state.current %>">
      <%= for _position <- @state.positions do %>
        <button></button>
      <% end %>
    </div>

We loop over all positions in the `@state` struct we assigned in the
live view and draw 81 empty `<button>` elements in a `<div>`, which the
stylesheet automatically styles as a Go board. We'll also add the
current color as a class name to the board, to allow the CSS to show the
stone you're about to place when you hover over a position.

Finally, we'll route any requests to `/` to our `GameLive` module in our
router.

    # lib/hayago_web/router.ex
    defmodule HayagoWeb.Router
      # ...

      scope "/", HayagoWeb do
        pipe_through :browser

        live "/", GameLive
      end
    end