Place stones on the board with State.place/2
To place stones on the board, `State` implements a `place/2` function
that takes a `State` struct, and an index. It will use that index to
replace the position that corresponds to the index with the value in the
`:current` key, which is either `:black` or `:white`, depending on which
player's turn it is.

In the template, we add a `phx-click` and `phx-value` attribute to our
buttons. These attributes tell LiveView to send an event to our
`GameLive` module.

    # lib/hayago_web/templates/game/index.html.leex
    <div class="board">
      <%= for {value, index} <- Enum.with_index(@state.positions) do %>
        <button phx-click="place" phx-value="<%= index %>" class="<%= value %>"></button>
      <% end %>
    </div>

In our live view, we hanfdle the event by matching on the attributes we
set in the template. We use the passed index to call `State.place/2`,
which returns a new state with the stone placed on the board.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_event("place", index, %{assigns: assigns} = socket) do
        new_state = State.place(assigns.state, String.to_integer(index))
        {:noreply, assign(socket, state: new_state)}
      end
    end