Add State to describe game states
In our game, the `State` module describes the state of the game. It
keeps track of which stones are on the board in its `:positions` list, and
it knows which player is up next in its `:current` key.

    # lib/hayago/state.ex
    defmodule Hayago.State do
      alias Hayago.State
      defstruct positions: Enum.map(1..81, fn _ -> nil end), current: :black

      # ...
    end

A new state automatically creates a list of 81 `nil` values as its
positions, as the board is empty and has 9 by 9 positions. The current
player is black, as the player with the black stones is the first to
move.

The `State` module exposes two functions. First is `place/2`, which
places a new stone on the board. If the new stone steals all of another
stone's liberties, the captured stone is removed from the board
automatically.

The `legal?/2` function checks if a move is legal by checking if the
position is already occupied by another stone, and trying the move using
the `place/2` function to make sure it's not immediately captured.