Disable illegal positions in index.html.leex
To prevent illegal moves, we'll render a disabled button for every
position that either has a stone on it already, or ones where a newly
placed stone would be captured immediately because it has no liberties.

To make that work, we'll use `State.legal/2`, which takes the current
state and an index and returns wether the current player can place a
stone there.

    <div class="board <%= @state.current %>">
      <%= for {value, index} <- Enum.with_index(@state.positions) do %>
        <%= if Hayago.State.legal?(@state, index) do %>
          <button phx-click="place" phx-value="<%= index %>" class="<%= value %>"></button>
        <% else %>
          <button class="<%= value %>" disabled="disabled"></button>
        <% end %>
      <% end %>
    </div>

Because LiveView takes care of updating the page, we can add an
if-statement to the template that checks if placing a stone on each
position is a legal move. If it is, we'll render the same button as
before. If not, we'll render a disabled button.