Show captures on index.leex
Whenever a stone is captured, a counter is incremented in `State`'s
`:captures` map, which holds a counter for both the black and the white
stones.

For each captured stone, we'll show a stone above the board. Since the
captures are already available in the `@state` struct we receive from
the live view, we'll loop over each of the counters to render a `<span>`
with the correct class name for the stylesheet to turn into a button.

    <div class="captures">
      <div>
        <%= for _ <- 1..@state.captures.black, @state.captures.black > 0 do %>
          <span class="black"></span>
        <% end %>
      </div>
      <div>
        <%= for _ <- 1..@state.captures.white, @state.captures.white > 0 do %>
          <span class="white"></span>
        <% end %>
      </div>
    </div>

Since we're using a range in our list comprehension, we'll make sure to
add a filter to make sure the list isn't empty when looping over it.