Add Heroku builpack and production configuration
As per https://hexdocs.pm/phoenix/heroku.html.

Create a new Heroku application with the [hashnuke/elixir
buildpack](https://github.com/HashNuke/heroku-buildpack-elixir).

    $ heroku create --buildpack hashnuke/elixir
    Creating app... done, ⬢ cryptic-harbor-83944
    Setting buildpack to hashnuke/elixir... done
    https://cryptic-harbor-83944.herokuapp.com/ | https://git.heroku.com/cryptic-harbor-83944.git

A new application is created with a random name. Run the `heroku apps:rename`
command to rename the application.

    $ heroku apps:rename hayago
    Renaming cryptic-harbor-83944 to hayago... done
    https://hayago.herokuapp.com/ | https://git.heroku.com/hayago.git
    Git remote heroku updated
    ▸    Don't forget to update git remotes for all other local checkouts of the app.
    In `config/prod.exs`, set the `:scheme

Add the Phoenix static buildpack to compile static assets.

    $ heroku buildpacks:add https://github.com/gjaldon/heroku-buildpack-phoenix-static.git
    Buildpack added. Next release on hayago will use:
      1. hashnuke/elixir
      2. https://github.com/gjaldon/heroku-buildpack-phoenix-static.git
    Run git push heroku master to create a new release using these buildpacks.

Set the `SECRET_KEY_BASE` environment variable on Heroku.

    $ heroku config:set SECRET_KEY_BASE="$(mix phx.gen.secret)"
    Setting SECRET_KEY_BASE and restarting ⬢ hayago... done, v3

    $ heroku config:set SECRET_KEY_BASE="$(mix phx.gen.secret)"
    Setting SECRET_KEY_BASE and restarting ⬢ hayago... done, v3
    SECRET_KEY_BASE: ==> phoenix_live_reload

Configure the application in `config/prod.exs` with Heroku's `SCHEME`
and `HOST_NAME` environment variables in the `:url` key, and
`[rewrite_on: [:x_forwarded_proto]]` as the `:force_ssl` key.

    # config/prod.exs
    config :hayago, HayagoWeb.Endpoint,
      url: [scheme: System.get_env("SCHEME"), host: System.get_env("HOST_NAME"), port: 443],
      force_ssl: [rewrite_on: [:x_forwarded_proto]],
      cache_static_manifest: "priv/static/cache_manifest.json"

Add a file named `elixir_buildpack.config` to set the Elixir and Erlang
versions the application should use.

    # elixir_buildpack.config
    elixir_version=1.8.2
    erlang_version=21.3.7

Deploy the app by pushing to the `heroku` remote.

    $ git push heroku master

Remember that when pushing from a local branch that is *not* called `master`,
the remote branch should be prefixed with the local branch name.

    $ git push heroku my_branch:master

Where `heroku` is the remote, `my_branch` is the local branch name, and
`master` is the remote branch name.