Add Game struct to maintain state history
To allow players to undo and redo moves, the game needs to keep a
history of the moves that are made since it started. Currently, it keeps
one `State` struct, which is accessible as `@state` in the live view.

    # lib/hayago_web/templates/game/index.html.leex
    <div class="board <%= @state.current %>">
      <%= for {value, index} <- Enum.with_index(@state.positions) do %>
        <%= if Hayago.State.legal?(@state, index) do %>
          <button phx-click="place" phx-value="<%= index %>" class="<%= value %>"></button>
        <% else %>
          <button class="<%= value %>" disabled="disabled"></button>
        <% end %>
      <% end %>
    </div>

Whenever a move is made by clicking one of the invisible buttons on the
board, the `GameLive` module handles the event. It uses `State.place/2`
to create a new state struct and assigns that as the new state to update
the view.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_event("place", index, %{assigns: assigns} = socket) do
        new_game = Game.place(assigns.game, String.to_integer(index))
        {:noreply, assign(socket, game: new_game, state: Game.state(new_game))}
      end
    end

Because of this implementation, there's currently no way to jump back in
history, as the state is replaced for each move.

To retain history, we'll add a struct named `Game`, which keeps a
history of states in its `:history` attribute. When the game starts, it
has a single, empty state in its history list, to represent the empty
board.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      alias Hayago.{Game, State}
      defstruct history: [%State{}]

      # ...
    end

We'll add a convenience function to get the current state from a game.
Since each newly added state is prepended to the history list, we can
take the list's head to get the current state.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...

      def state(%Game{history: [state | _]}) do
        state
      end

      # ...
    end

Finally, we'll implement a `place/2` function to the `Game` module. It
uses `State.place/2` to create a new state struct, and prepends that to
the history list.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...

      def place(%Game{history: [state | _] = history} = game, position) do
        %{game | history: [State.place(state, position) | history]}
      end
    end

Now, let's use the new game struct in the `GameLive` module. The
`mount/2` function is used to set up the game. Instead of creating a
state struct and assigning it directly to the socket, we'll create a new
game struct, and assign that to the `:game` assign.

The `:state` assign is taken from the game struct for convenience, to
prevent having to call to `Game.state/1` from the view or template.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      alias Hayago.Game
      use Phoenix.LiveView

      # ...

      def mount(_session, socket) do
        game = %Game{}
        {:ok, assign(socket, game: game, state: Game.state(game))}
      end

      # ...
    end

When placing a stone, we'll now call the `Game.place/2` function to
update the current state while retaining the history list. Again, we use
`Game.state/1` on the updated game to preload the state.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      alias Hayago.Game
      use Phoenix.LiveView

      # ...

      def handle_event("place", index, %{assigns: assigns} = socket) do
        new_game = Game.place(assigns.game, String.to_integer(index))
        {:noreply, assign(socket, game: new_game, state: Game.state(new_game))}
      end
    end

If we try our game again, we'll notice nothing visible has changed.
While we keep history of all moves, we don't currently do anything with
the history.