Add :index attribute to Game
Although we're now keeping a history of game states, all functions in
out `Game` module still only use the first state in the list. As each
new state is prepended to the history list, the first in the list is
always the latest state.

To allow jumping back and forward through we'll add an attribute named
`:index` to the `Game` struct, which defaults to 0.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      alias Hayago.{Game, State}
      defstruct history: [%State{}], index: 0

      # ...
    end

By having an index, we can jump back to a previous state, without having
to drop states from the front of the list. Instead, we'll update our
`state/1` function to take the index into account

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...

      def state(%Game{history: history, index: index}) do
        Enum.at(history, index)
      end

      # ...
    end

Now, if we have a game with a three-state history, setting the `:index`
attribute to 1 will return the second state in the list, essentially
reverting back to that state.