Add undo and redo buttons
Now that our game module keeps a history of moves, and we can jump
between them, let's add buttons to undo and redo moves while playing the
game.

    # lib/hayago_web/templates/game/index.html.leex
    # ...

    <div class="history">
      <button phx-click="jump" phx-value="<%= @game.index + 1%>">Undo</button>
      <button phx-click="jump" phx-value="<%= @game.index - 1%>">Redo</button>
    </div>

First, we'll add two more buttons. We'll use "jump" as the value for
both buttons' `phx-click` attribute, which is the name for a function
we'll implement in the `GameLive` module shortly.

The `phx-value` attribute is used to pass the history index we'd like to
jump to. The history list is reversed as new moves are prepended to the
history list. So to undo a move, we'll increase the current index by 1,
and we'll decrease it by 1 to redo.

In the `GameLive` module, we'll handle the jump event by calling
`Game.jump/2` with the current game and the passed index to get a new
game struct we'll assign on the socket. Like before, we update the state
assign for convenience.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_event("jump", destination, %{assigns: %{game: game}} = socket) do
        new_game = Game.jump(game, String.to_integer(destination))
        {:noreply, assign(socket, game: new_game, state: Game.state(new_game))}
      end
    end