Slice history before prepending new states
If we open our browser and navigate to https://localhost:4000, we can
see that the undo and redo buttons work. After placing a few stones, we
can click the undo button to get the last one removed, and the redo
button to get it back again.

However, if we undo a move and try to place a stone, we notice that the
new stone doesn't get added to the board. Instead, the stone we just
removed by pressing the undo button reappears.

It turns out we have one more step to take before we can place a new
stone after pressing the undo button. Let's break down what's happening.

1. We place a stone on the board, which prepends a new state to the
   history list.
2. We press the undo button to increase the history index to 1,
   bringing us back to our initial state, giving us an empty board
   again.
3. We try to place a new stone, which prepends a new state to the
   history list.

The new state is the first in the list, meaning its index is 0, but the
Game index is still 1. That index belongs to the move we've undone by
pressing the undo button. Essentially, each of our newly added stones
are delayed by one move.

To fix this, we need to slice any undone states from the list whenever we
add a new state by placing a new stone.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...

      def place(%Game{history: history, index: index} = game, position) do
        new_state =
          game
          |> Game.state()
          |> State.place(position)

        %{game | history: [new_state | Enum.slice(history, index..-1)], index: 0}
      end

      # ...
    end

In the new version of our `place/2` function, we use `Enum.slice/2` to
drop the undone moves from the history list. We'll also reset our game's
index attribute to 0, which makes sure our newly added stones
always immediately appear.