Disable undo and redo buttons when not applicable
Finally, we need to make sure the undo and redo buttons are disabled
when there are no moves to undo or redo. To do that, we'll implement a
function named `history?/2`, which takes a game struct and an index and
returns wether that index is part of the game's history.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...

      def history?(%Game{history: history}, index) when index >= 0 and length(history) > index do
        true
      end

      def history?(_game, _index), do: false
    end

Our function checks if the requested index is above 0 to make sure the
game doesn't allow jumping forward when there are no moves done yet. Then,
it checks if the length of the history list is higher than the index, to
make sure we don't overshoot the list when undoing moves.

With our new function in place, we can check if the game can undo and
redo to the previous and next move before rendering the button. If it
can't it will render a disabled version.

    # lib/hayago_web/templates/game/index.html.leex
    <div class="history">
      <%= if Hayago.Game.history?(@game, @game.index + 1) do %>
        <button phx-click="jump" phx-value="<%= @game.index + 1%>">Undo</button>
      <% else %>
        <button disabled="disabled">Undo</button>
      <% end %>

      <%= if Hayago.Game.history?(@game, @game.index - 1) do %>
        <button phx-click="jump" phx-value="<%= @game.index - 1%>">Redo</button>
      <% else %>
        <button disabled="disabled">Redo</button>
      <% end %>
    </div>

Trying our game again now, we'll see that we can't undo further than the
list of previous moves, and we can't redo into the future.