Update Phoenix to 1.4.8, phoenix_live_view to 957b7e
To allow multiple players to connect to the same game, we'll add the
game's name to an URL the first user can share. When a user starts a new
game, we'll use the pushState function from the HTML5 history API to add
the game's name to the URL without reloading the page.

We'll use Phoenix LiveView's `live_redirect/2` function for that, which
was added recently, after we started working on our game. To make sure
we're on a recent enough version, we'll update the dependency before we
continue.

    $ mix deps.update phoenix_live_view