Implement Game GenServer callbacks, add Supervisor
Currently, the game's state is kept in the player's socket connection.
This works for games with a single player connection, but there's no way
for another socket connection to access an existing game. To allow two
players to play the game together over two socket connections, we need
to store each started game in a seperate process.

We'll use a dynamically supervised GenServer to keep each game's state.
The dynamic supervisor allows players to start games and it
automatically restart them when a game process crashes.

First, let's turn our Game module into a GenServer.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...
      use GenServer

      def start_link(options) do
        GenServer.start_link(__MODULE__, %Game{}, options)
      end

      @impl true
      def init(game) do
        {:ok, game}
      end

      @impl true
      def handle_call(:game, _from, game) do
        {:reply, game, game}
      end

      @impl true
      def handle_cast({:place, position}, game) do
        {:noreply, Game.place(game, position)}
      end

      @impl true
      def handle_cast({:jump, destination}, game) do
        {:noreply, Game.jump(game, destination)}
      end

      # ...
    end

Our GenServer handles the `:game` call, which returns the current game's
`Game` struct. The `{:place, position}` and `{:jump, destination}`
update the game by calling `Game.place/2` and `Game.jump/2`,
respectively.

With our GenServer callbacks in place, we can spawn a process that keeps
a game's state. To supervise these processes, we'll use
`DynamicSupervisor`, which allows spawning children on demand. We'll use
it to start a new game when a player opens the page.

    # lib/hayago/application.ex
    defmodule Hayago.Application do
      # ...

      def start(_type, _args) do
        # List all child processes to be supervised
        children = [
          HayagoWeb.Endpoint,
          {DynamicSupervisor, strategy: :one_for_one, name: Hayago.GameSupervisor}
        ]

        opts = [strategy: :one_for_one, name: Hayago.Supervisor]
        Supervisor.start_link(children, opts)
      end

      # ...
    end

We'll set up our dynamic supervisor in our app's `Application` module.
The `:one_for_one` strategy (which is the only one available for
dynamic supervisors) makes sure to restart a game whenever it crashes.