Add Registry to application supervisor
Because our Game module now implements GenServer callbacks, we can spawn
a Game through the GenServer module.

    % iex -S mix
    iex(1)> {:ok, pid} = DynamicSupervisor.start_child(Hayago.GameSupervisor, Hayago.Game)
    {:ok, #PID<0.286.0>}
    iex(2)> GenServer.cast(pid, {:place, 0})
    :ok

The returned pid is used to get and update the game's state. In this
case, we use it to place a stone on the top left position on the board.

Instead of keeping the game in the socket for every connection, we could
add the pid to the socket and request and update its state through that.

However, our supervisor is tasked with restarting any game process that
crashes, which results in a new process, with a new pid. If that
happens, our players are still disconnected from game because the only
reference they have to the game is no longer working, and they have no
way to get the new pid.

    iex(3)> Process.exit(pid, :kill)
    true
    iex(4)> Process.alive?(pid)
    false

A solution to this is naming processes when they're spawned to refer to
them by name. Whenever a named process is restarted by a supervisor, the
newly started process automatically receives the name of the process it
replaces.

    iex(5)> {:ok, pid} = DynamicSupervisor.start_child(Hayago.GameSupervisor, {Hayago.Game, name: :game_1})
    {:ok, #PID<0.285.0>}
    iex(6)> Process.whereis(:game_1)
    #PID<0.285.0>
    iex(7)> Process.exit(pid, :kill)
    true
    iex(8)> Process.whereis(:game_1)
    #PID<0.288.0>
    iex(9)> GenServer.cast(pid, {:place, 0})
    :ok

Using a process' name instead of its pid, we can refer to it, even if it
was restarted by the supervisor. We'll need to generate a unique name
for each process when spawning it, and keep that in our socket to refer
to the game later.

But, wait. We're naming our processes with atoms here. Since atoms
aren't garbage collected, spawning a lot of games can exhaust our
memory. Instead, we'd like to use strings, which *are* garbage
collected.

Because we can't name a process with a string, we need to use a Registry
to link string names to game pids. Elixir's GenServer implementation has
a built-in way to refer to processes in a registry through it's
`:via`-tuples.

First, let's start the registry in our main supervisor whenever our
application starts.

    # lib/hayago/application.ex
    defmodule Hayago.Application do
      # ...

      def start(_type, _args) do
        children = [
          HayagoWeb.Endpoint,
          {Registry, keys: :unique, name: Hayago.GameRegistry},
          {DynamicSupervisor, strategy: :one_for_one, name: Hayago.GameSupervisor}
        ]

        opts = [strategy: :one_for_one, name: Hayago.Supervisor]
        Supervisor.start_link(children, opts)
      end

      # ...
    end

With our new registry in place, we can start processes using a string as
their names, and refer to them later without creating an atom per
spawned game.

    iex(1)> {:ok, pid} = DynamicSupervisor.start_child(Hayago.GameSupervisor, {Hayago.Game, name: {:via, Registry, {Hayago.GameRegistry, "game_1"}}})
    {:ok, #PID<0.294.0>}
    iex(2)> Process.exit(pid, :kill)
    iex(3)> GenServer.cast({:via, Registry, {Hayago.GameRegistry, "game_1"}}, {:place, 0})
    :ok