Use dynamically supervised games in GameLive
Instead of creating a new game struct and assigning it directly to the
socket, we'll update the `GameLive.mount/2` function to start a
supervised process that holds a game's state.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def mount(_session, socket) do
        name =
          ?a..?z
          |> Enum.take_random(6)
          |> List.to_string()

        {:ok, _pid} =
          DynamicSupervisor.start_child(Hayago.GameSupervisor, {Game, name: via_tuple(name)})

        {:ok, assign_game(socket, name)}
      end

      # ...

      defp via_tuple(name) do
        {:via, Registry, {Hayago.GameRegistry, name}}
      end

      defp assign_game(socket, name) do
        socket
        |> assign(name: name)
        |> assign_game()
      end

      defp assign_game(%{assigns: %{name: name}} = socket) do
        game = GenServer.call(via_tuple(name), :game)
        assign(socket, game: game, state: Game.state(game))
      end
    end

We'll create a random six character string as the name for our process,
which we'll use when spawning the process through our `GameSupervisor`.
The `via_tuple/1` convenience function returns the `:via`-tuple we'll
need to register the process' name in the registry.

Finally, we addd an `assign_game/1-2` function that takes a socket the
game's registered name. It calls out to the GenServer process to get the
game's current state before assigning it to the socket.

Then, we'll switch both `handle_event/3` variants to use the game via
the GenServer.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_event("place", index, %{assigns: %{name: name}} = socket) do
        :ok = GenServer.cast(via_tuple(name), {:place, String.to_integer(index)})
        {:noreply, assign_game(socket)}
      end

      def handle_event("jump", destination, %{assigns: %{name: name}} = socket) do
        :ok = GenServer.cast(via_tuple(name), {:jump, String.to_integer(destination)})
        {:noreply, assign_game(socket)}
      end

      # ...
    end

We'll use the `via_tuple/1` function again to cast both our `:place` and
`:jump` functions directly on the game's process. We'll then reassign
the game to the socket by calling our `assign_game/1` function.