Add Game name to parameters with live_redirect/2
When a player visits our game without a game name in the URL, our app
will start a new one and update the URL to include the newly created
game's name. In our case, the game is served on the root of our
application, so visiting http://localhost:4000 will redirect to
http://localhost:4000?name=abcdef, where "abcdef" is our game's name.

To do this, we'll replace our `mount/2` function with two variants of
the `handle_params/3` function.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_params(%{"name" => name} = _params, _uri, socket) do
        {:noreply, assign_game(socket, name)}
      end

      def handle_params(_params, _uri, socket) do
        name =
          ?a..?z
          |> Enum.take_random(6)
          |> List.to_string()

        {:ok, _pid} =
          DynamicSupervisor.start_child(Hayago.GameSupervisor, {Game, name: via_tuple(name)})

        {:ok,
        live_redirect(
          socket,
          to: HayagoWeb.Router.Helpers.live_path(socket, HayagoWeb.GameLive, name: name)
        )}
      end

      # ...
    end

The first variant handles requests that have a name in its URL
parameters. In that case, the latest game state is fetched from the
process corresponding to the name from the parameters and assigned to
the socket using the `assign_game/2` function.

The second is a lot like our `mount/2` function we're replacing. We're
generating a name and starting a game. However, instead of assigning the
game to the socket and returning, this function uses LiveView's
`live_redirect/2` function to add the name to the current URL. After the
URL changes, the first variant is automatically executed, assigning the
name and game state to the socket.