Add PubSub connections to broadcast game updates
If we open our game right now, we can see that every visit to
http://localhost:4000 gets redirected to a URL with a name query
parameter. Opening that URL twice connects two sockets to the same game.

However, placing a stone in one of the windows, it doesn't yet appear in
the other one, although it shows the correct stones when refreshing the
window.

Everything is wired up correctly, but the second socket doesn't get
notified when the first makes a move yet. We need a way to get all
connected sockets to subscribe to a system that publishes updates to all
clients when one makes a move.

We'll use `Phoenix.PubSub` to make that happen. Back in our `GameLive`
module, we'll subscribe each socket connection to a channel when a game
is created, and we'll broadcast a message over that channel whenever we
place a stone or travel in history.

    # lib/hayago_web/live/game_live.ex
    defmodule HayagoWeb.GameLive do
      # ...

      def handle_params(%{"name" => name} = _params, _uri, socket) do
        :ok = Phoenix.PubSub.subscribe(Hayago.PubSub, name)
        {:noreply, assign_game(socket, name)}
      end

      # ...

      def handle_event("place", index, %{assigns: %{name: name}} = socket) do
        :ok = GenServer.cast(via_tuple(name), {:place, String.to_integer(index)})
        :ok = Phoenix.PubSub.broadcast(Hayago.PubSub, name, :update)
        {:noreply, assign_game(socket)}
      end

      def handle_event("jump", destination, %{assigns: %{name: name}} = socket) do
        :ok = GenServer.cast(via_tuple(name), {:jump, String.to_integer(destination)})
        :ok = Phoenix.PubSub.broadcast(Hayago.PubSub, name, :update)
        {:noreply, assign_game(socket)}
      end

      def handle_info(:update, socket) do
        {:noreply, assign_game(socket)}
      end

      # ...
    end

In `handle_params/3` (the variant that matches on URLs with name query
parameters), we'll call out to `Phoenix.PubSub.subscribe/2`. We'll use
the game's name as the channel's topic. Because all connected clients
hit this function, we'll know they'll all be subscribed to this topic.

In both `handle_event/3` functions, we'll use
`Phoenix.PubSub.broadcast/3` to send a message to all subscribers in the
topic. We'll send `:update` as the message, which is then picked up in a
newly added `handle_info/2` function, which fetches the current Game's
state and updates the view whenever it receives an update message.