Add one-hour Game timeout
To make sure games are cleaned up, we need to terminate their processes
when they're no longer being used. For now, we'll add a timeout to the
GenServer's callback functions. When a game hasn't had any interaction
for ten minutes, the process will send itself a `:timeout` message to
stop the process.

    # lib/hayago/game.ex
    defmodule Hayago.Game do
      # ...
      use GenServer, restart: :transient

      @timeout 600_000

      def start_link(options) do
        GenServer.start_link(__MODULE__, %Game{}, options)
      end

      @impl true
      def init(game) do
        {:ok, game, @timeout}
      end

      @impl true
      def handle_call(:game, _from, game) do
        {:reply, game, game, @timeout}
      end

      @impl true
      def handle_cast({:place, position}, game) do
        {:noreply, Game.place(game, position), @timeout}
      end

      @impl true
      def handle_cast({:jump, destination}, game) do
        {:noreply, Game.jump(game, destination), @timeout}
      end

      @impl true
      def handle_info(:timeout, game) do
        {:stop, :normal, game}
      end

      # ...
    end

By adding a timeout in milliseconds to every callback's response tuple,
we tell the GenServer to send itself a timeout message when the process
hasn't received a message for ten minutes. The `:timeout` callback
returns a `:stop`-tuple to tell the process to stop.

We'll also make sure to set the `:restart` value to `:transient` when
including the GenServer code to make sure the supervisor only restarts
the game when it terminates abnormally.